reorganized network
