+++ src/main/java/network/data/SharedData.java
@@ -1,4 +1,4 @@
-package network;
+package network.data;
 
 import java.util.ArrayList;
 import java.util.concurrent.BlockingQueue;
